+++ spec/helpers/bundler_wrapper_spec.rb
@@ -1,7 +1,18 @@
 require 'spec_helper'
 
-describe "BundlerWrapper" do
+describe "Bundle platform conversion" do
+  it "converts `bundle platform --ruby` for prerelease versions" do
+    actual = LanguagePack::Helpers::BundlerWrapper.platform_to_version("ruby 3.3.0.preview2")
+    expect(actual).to eq("ruby-3.3.0.preview2")
+  end
 
+  it "converts `bundle platform --ruby` for released versions" do
+    actual = LanguagePack::Helpers::BundlerWrapper.platform_to_version("ruby 3.1.4")
+    expect(actual).to eq("ruby-3.1.4")
+  end
+end
+
+describe "BundlerWrapper" do
   before(:each) do
     if ENV['RUBYOPT']
       @original_rubyopt = ENV['RUBYOPT']